fix mot modeling
