+++ ppdet/modeling/architectures/jde.py
@@ -17,7 +17,6 @@ from __future__ import division
 from __future__ import print_function
 
 import paddle
-import numpy as np
 from ppdet.core.workspace import register, create
 from .meta_arch import BaseArch
 
@@ -26,6 +25,14 @@ __all__ = ['JDE']
 
 @register
 class JDE(BaseArch):
+    """
+    JDE network, see https://arxiv.org/abs/1909.12605v1
+
+    Args:
+        detector (object): detector model instance
+        reid (object): reid model instance
+        tracker (object): tracker instance
+    """
     __category__ = 'architecture'
 
     def __init__(self,
@@ -61,8 +68,8 @@ class JDE(BaseArch):
 
         if self.training:
             emb_feats = det_outs['emb_feats']
-            loss_confs = det_outs['yolo_losses']['loss_confs']
-            loss_boxes = det_outs['yolo_losses']['loss_boxes']
+            loss_confs = det_outs['det_losses']['loss_confs']
+            loss_boxes = det_outs['det_losses']['loss_boxes']
             jde_losses = self.reid(emb_feats, self.inputs, loss_confs,
                                    loss_boxes)
             return jde_losses
@@ -80,17 +87,14 @@ class JDE(BaseArch):
                 bbox = det_outs['bbox']
                 nms_keep_idx = det_outs['nms_keep_idx']
 
-                emb_valid = paddle.gather_nd(emb_outs, boxes_idx)
-                embeddings = paddle.gather_nd(emb_valid, nms_keep_idx)
+                pred_dets = paddle.concat((bbox[:, 2:], bbox[:, 1:2]), axis=1)
 
-                pred_boxes = bbox[:, 2:].numpy()
-                pred_scores = bbox[:, 1:2].numpy()
-                pred_dets = np.concatenate((pred_boxes, pred_scores), axis=1)
-                pred_embs = embeddings.numpy()
-                img0_shape = self.inputs['img0_shape'].numpy()[0]
+                emb_valid = paddle.gather_nd(emb_outs, boxes_idx)
+                pred_embs = paddle.gather_nd(emb_valid, nms_keep_idx)
+                scale_factor = self.inputs['scale_factor']
 
                 online_targets = self.tracker.update(pred_dets, pred_embs,
-                                                     img0_shape)
+                                                     scale_factor)
                 return online_targets
 
             else:
