+++ ppdet/modeling/mot/tracker/deepsort_tracker.py
@@ -13,10 +13,10 @@
 # limitations under the License.
 
 import numpy as np
-from collections import OrderedDict
 
-from ..matching import iou_matching, linear_assignment
-from ..matching.nn_matching import NearestNeighborDistanceMetric
+from ..matching.deepsort_matching import NearestNeighborDistanceMetric
+from ..matching.deepsort_matching import iou_cost, min_cost_matching, matching_cascade, gate_cost_matrix
+from .base_sde_tracker import Track
 
 from ppdet.core.workspace import register, serializable
 from ppdet.utils.logger import setup_logger
@@ -25,240 +25,81 @@ logger = setup_logger(__name__)
 __all__ = ['DeepSORTTracker']
 
 
-class TrackState(object):
-    """
-    Enumeration type for the single target track state. Newly created tracks are
-    classified as `tentative` until enough evidence has been collected. Then,
-    the track state is changed to `confirmed`. Tracks that are no longer alive
-    are classified as `deleted` to mark them for removal from the set of active
-    tracks.
-    """
-    Tentative = 1
-    Confirmed = 2
-    Deleted = 3
-
-
-class Track(object):
-    """
-    A single target track with state space `(x, y, a, h)` and associated
-    velocities, where `(x, y)` is the center of the bounding box, `a` is the
-    aspect ratio and `h` is the height.
-
-    Parameters
-    ----------
-    mean : ndarray
-        Mean vector of the initial state distribution.
-    covariance : ndarray
-        Covariance matrix of the initial state distribution.
-    track_id : int
-        A unique track identifier.
-    n_init : int
-        Number of consecutive detections before the track is confirmed. The
-        track state is set to `Deleted` if a miss occurs within the first
-        `n_init` frames.
-    max_age : int
-        The maximum number of consecutive misses before the track state is
-        set to `Deleted`.
-    feature : Optional[ndarray]
-        Feature vector of the detection this track originates from. If not None,
-        this feature is added to the `features` cache.
-
-    Attributes
-    ----------
-    mean : ndarray
-        Mean vector of the initial state distribution.
-    covariance : ndarray
-        Covariance matrix of the initial state distribution.
-    track_id : int
-        A unique track identifier.
-    hits : int
-        Total number of measurement updates.
-    age : int
-        Total number of frames since first occurance.
-    time_since_update : int
-        Total number of frames since last measurement update.
-    state : TrackState
-        The current track state.
-    features : List[ndarray]
-        A cache of features. On each measurement update, the associated feature
-        vector is added to this list.
-    """
-
-    def __init__(self,
-                 mean,
-                 covariance,
-                 track_id,
-                 n_init,
-                 max_age,
-                 feature=None):
-        self.mean = mean
-        self.covariance = covariance
-        self.track_id = track_id
-        self.hits = 1
-        self.age = 1
-        self.time_since_update = 0
-
-        self.state = TrackState.Tentative
-        self.features = []
-        if feature is not None:
-            self.features.append(feature)
-
-        self._n_init = n_init
-        self._max_age = max_age
-
-    def to_tlwh(self):
-        """Get position in format `(top left x, top left y, width, height)`."""
-        ret = self.mean[:4].copy()
-        ret[2] *= ret[3]
-        ret[:2] -= ret[2:] / 2
-        return ret
-
-    def to_tlbr(self):
-        """Get position in bounding box format `(min x, miny, max x, max y)`."""
-        ret = self.to_tlwh()
-        ret[2:] = ret[:2] + ret[2:]
-        return ret
-
-    def predict(self, kalman_filter):
-        """Propagate the state distribution to the current time step using a
-        Kalman filter prediction step.
-        """
-        self.mean, self.covariance = kalman_filter.predict(self.mean,
-                                                           self.covariance)
-        self.age += 1
-        self.time_since_update += 1
-
-    def update(self, kalman_filter, detection):
-        """Perform Kalman filter measurement update step and update the associated
-        detection feature cache.
-        """
-        self.mean, self.covariance = kalman_filter.update(self.mean,
-                                                          self.covariance,
-                                                          detection.to_xyah())
-        self.features.append(detection.feature)
-
-        self.hits += 1
-        self.time_since_update = 0
-        if self.state == TrackState.Tentative and self.hits >= self._n_init:
-            self.state = TrackState.Confirmed
-
-    def mark_missed(self):
-        """Mark this track as missed (no association at the current time step).
-        """
-        if self.state == TrackState.Tentative:
-            self.state = TrackState.Deleted
-        elif self.time_since_update > self._max_age:
-            self.state = TrackState.Deleted
-
-    def is_tentative(self):
-        """Returns True if this track is tentative (unconfirmed)."""
-        return self.state == TrackState.Tentative
-
-    def is_confirmed(self):
-        """Returns True if this track is confirmed."""
-        return self.state == TrackState.Confirmed
-
-    def is_deleted(self):
-        """Returns True if this track is dead and should be deleted."""
-        return self.state == TrackState.Deleted
-
-
 @register
 @serializable
 class DeepSORTTracker(object):
     __inject__ = ['motion']
     """
-    DeepSORT tracker.
+    DeepSORT tracker
 
-    Parameters
-    ----------
-    img_size: array_like
-        input shape in format '[w, h]'.
-    metric_type : str
-        Either "euclidean" or "cosine".
-    matching_threshold: float
-        The matching threshold. Samples with larger distance are considered an
-        invalid match.
-    budget : Optional[int]
-        If not None, fix samples per class to at most this number. Removes
-        the oldest samples when the budget is reached.
-    max_age : int
-        Maximum number of missed misses before a track is deleted.
-    n_init : int
-        Number of consecutive detections before the track is confirmed. The
-        track state is set to `Deleted` if a miss occurs within the first
-        `n_init` frames.
-
-    Attributes
-    ----------
-    metric : nn_matching.NearestNeighborDistanceMetric
-        The distance metric used for measurement to track association.
-    max_age : int
-        Maximum number of missed misses before a track is deleted.
-    n_init : int
-        Number of frames that a track remains in initialization phase.
-    kf : kalman_filter.KalmanFilter
-        A Kalman filter to filter target trajectories in image space.
-    tracks : List[Track]
-        The list of active tracks at the current time step.
+    Args:
+        img_size (list): input image size, [h, w]
+        budget (int): If not None, fix samples per class to at most this number.
+            Removes the oldest samples when the budget is reached.
+        max_age (int): maximum number of missed misses before a track is deleted
+        n_init (float): Number of frames that a track remains in initialization
+            phase. Number of consecutive detections before the track is confirmed. 
+            The track state is set to `Deleted` if a miss occurs within the first 
+            `n_init` frames.
+        metric_type (str): either "euclidean" or "cosine", the distance metric 
+            used for measurement to track association.
+        matching_threshold (float): samples with larger distance are 
+            considered an invalid match.
+        max_iou_distance (float): max iou distance threshold
+        motion (object): KalmanFilter instance
     """
 
     def __init__(self,
-                 img_size=[1088, 608],
-                 metric_type='cosine',
-                 matching_threshold=0.2,
+                 img_size=[608, 1088],
                  budget=100,
-                 max_iou_distance=0.7,
                  max_age=30,
                  n_init=3,
+                 metric_type='cosine',
+                 matching_threshold=0.2,
+                 max_iou_distance=0.7,
                  motion='KalmanFilter'):
         self.img_size = img_size
+        self.max_age = max_age
+        self.n_init = n_init
         self.metric = NearestNeighborDistanceMetric(metric_type,
                                                     matching_threshold, budget)
         self.max_iou_distance = max_iou_distance
-        self.max_age = max_age
-        self.n_init = n_init
+        self.motion = motion
 
-        self.kf = motion
         self.tracks = []
         self._next_id = 1
 
     def predict(self):
-        """Propagate track state distributions one time step forward.
-
+        """
+        Propagate track state distributions one time step forward.
         This function should be called once every time step, before `update`.
         """
         for track in self.tracks:
-            track.predict(self.kf)
+            track.predict(self.motion)
 
     def update(self, detections):
-        """Perform measurement update and track management.
-
-        Parameters
-        ----------
-        detections : List[deep_sort.detection.Detection]
+        """
+        Perform measurement update and track management.
+        Args:
+            detections (list): List[ppdet.modeling.mot.utils.Detection]
             A list of detections at the current time step.
-
         """
         # Run matching cascade.
         matches, unmatched_tracks, unmatched_detections = \
             self._match(detections)
-        #print(11)
+
         # Update track set.
         for track_idx, detection_idx in matches:
-            self.tracks[track_idx].update(self.kf, detections[detection_idx])
-        #print(12)      
+            self.tracks[track_idx].update(self.motion,
+                                          detections[detection_idx])
         for track_idx in unmatched_tracks:
             self.tracks[track_idx].mark_missed()
-        #print(13)
         for detection_idx in unmatched_detections:
             self._initiate_track(detections[detection_idx])
-        #print(14)
         self.tracks = [t for t in self.tracks if not t.is_deleted()]
-        #print(15)
+
         # Update distance metric.
         active_targets = [t.track_id for t in self.tracks if t.is_confirmed()]
-        #print(16)
         features, targets = [], []
         for track in self.tracks:
             if not track.is_confirmed():
@@ -266,10 +107,8 @@ class DeepSORTTracker(object):
             features += track.features
             targets += [track.track_id for _ in track.features]
             track.features = []
-        #print(17)
         self.metric.partial_fit(
             np.asarray(features), np.asarray(targets), active_targets)
-        #print(18)
         output_stracks = self.tracks
         return output_stracks
 
@@ -278,10 +117,9 @@ class DeepSORTTracker(object):
             features = np.array([dets[i].feature for i in detection_indices])
             targets = np.array([tracks[i].track_id for i in track_indices])
             cost_matrix = self.metric.distance(features, targets)
-            cost_matrix = linear_assignment.gate_cost_matrix(
-                self.kf, cost_matrix, tracks, dets, track_indices,
-                detection_indices)
-
+            cost_matrix = gate_cost_matrix(self.motion, cost_matrix, tracks,
+                                           dets, track_indices,
+                                           detection_indices)
             return cost_matrix
 
         # Split track set into confirmed and unconfirmed tracks.
@@ -294,7 +132,7 @@ class DeepSORTTracker(object):
 
         # Associate confirmed tracks using appearance features.
         matches_a, unmatched_tracks_a, unmatched_detections = \
-            linear_assignment.matching_cascade(
+            matching_cascade(
                 gated_metric, self.metric.matching_threshold, self.max_age,
                 self.tracks, detections, confirmed_tracks)
 
@@ -308,8 +146,8 @@ class DeepSORTTracker(object):
             if self.tracks[k].time_since_update != 1
         ]
         matches_b, unmatched_tracks_b, unmatched_detections = \
-            linear_assignment.min_cost_matching(
-                iou_matching.iou_cost, self.max_iou_distance, self.tracks,
+            min_cost_matching(
+                iou_cost, self.max_iou_distance, self.tracks,
                 detections, iou_track_candidates, unmatched_detections)
 
         matches = matches_a + matches_b
@@ -317,7 +155,7 @@ class DeepSORTTracker(object):
         return matches, unmatched_tracks, unmatched_detections
 
     def _initiate_track(self, detection):
-        mean, covariance = self.kf.initiate(detection.to_xyah())
+        mean, covariance = self.motion.initiate(detection.to_xyah())
         self.tracks.append(
             Track(mean, covariance, self._next_id, self.n_init, self.max_age,
                   detection.feature))
