+++ ppdet/modeling/mot/utils.py
@@ -14,8 +14,9 @@
 
 import os
 import cv2
-import numpy as np
 import time
+import paddle
+import numpy as np
 
 __all__ = [
     'Timer',
@@ -75,9 +76,6 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, feature):
-        #self.tlwh = tlwh
-        #self.confidence = confidence
-
         self.tlwh = np.asarray(tlwh, dtype=np.float32)
         self.confidence = np.asarray(confidence, dtype=np.float32)
         self.feature = feature.numpy()
@@ -115,6 +113,50 @@ def load_det_results(det_file, num_frames):
     return results_list
 
 
+def scale_coords(coords, img_size, scale_factor):
+    scale_factor = scale_factor.numpy()[0]
+    ratio_h, ratio_w = scale_factor[0], scale_factor[1]
+    img0_shape = [int(img_size[0] / ratio_h), int(img_size[1] / ratio_w)]
+
+    ratio = min(ratio_h, ratio_w)
+    pad_w = (img_size[1] - img0_shape[1] * ratio) / 2
+    pad_h = (img_size[0] - img0_shape[0] * ratio) / 2
+    coords[:, 0::2] -= pad_w
+    coords[:, 1::2] -= pad_h
+    coords[:, 0:4] /= paddle.to_tensor(ratio)
+    coords[:, :4] = paddle.clip(coords[:, :4], min=0, max=coords[:, :4].max())
+    return coords.round()
+
+
+def clip_box(xyxy, img_size, scale_factor):
+    scale_factor = scale_factor.numpy()[0]
+    ratio_h, ratio_w = scale_factor[0], scale_factor[1]
+    img0_shape = [int(img_size[0] / ratio_h), int(img_size[1] / ratio_w)]
+
+    xyxy[:, 0::2] = paddle.clip(xyxy[:, 0::2], min=0, max=img0_shape[1])
+    xyxy[:, 1::2] = paddle.clip(xyxy[:, 1::2], min=0, max=img0_shape[0])
+    return xyxy
+
+
+def get_crops(xyxy, ori_img, pred_scores, w, h):
+    crops = []
+    keep_scores = []
+    xyxy = xyxy.numpy().astype(np.int64)
+    ori_img = ori_img.numpy()
+    ori_img = np.squeeze(ori_img, axis=0).transpose(1, 0, 2)
+    pred_scores = pred_scores.numpy()
+    for i, bbox in enumerate(xyxy):
+        if bbox[2] <= bbox[0] or bbox[3] <= bbox[1]:
+            continue
+        crop = ori_img[bbox[0]:bbox[2], bbox[1]:bbox[3], :]
+        crops.append(crop)
+        keep_scores.append(pred_scores[i])
+    if len(crops) == 0:
+        return [], []
+    crops = preprocess_reid(crops, w, h)
+    return crops, keep_scores
+
+
 def preprocess_reid(imgs, w, h):
     im_batch = []
     for img in imgs:
@@ -130,40 +172,3 @@ def preprocess_reid(imgs, w, h):
         im_batch.append(img)
     im_batch = np.concatenate(im_batch, 0)
     return im_batch
-
-
-def get_crops(xyxy, ori_img, pred_scores, w, h):
-    crops = []
-    keep_scores = []
-    xyxy = xyxy.astype(np.int64)
-    ori_img = np.squeeze(ori_img, axis=0).transpose(1, 0, 2)
-    for i, bbox in enumerate(xyxy):
-        if bbox[2] <= bbox[0] or bbox[3] <= bbox[1]:
-            continue
-        crop = ori_img[bbox[0]:bbox[2], bbox[1]:bbox[3], :]
-        crops.append(crop)
-        keep_scores.append(pred_scores[i])
-    if len(crops) == 0:
-        return [], []
-    crops = preprocess_reid(crops, w, h)
-    return crops, keep_scores
-
-
-def scale_coords(img_size, coords, img0_shape):
-    # img_size [w,h], img0_shape [h,w]
-    gain_w = float(img_size[0]) / img0_shape[1]  # gain  = old / new
-    gain_h = float(img_size[1]) / img0_shape[0]
-    gain = min(gain_w, gain_h)
-    pad_x = (img_size[0] - img0_shape[1] * gain) / 2  # width padding
-    pad_y = (img_size[1] - img0_shape[0] * gain) / 2  # height padding
-    coords[:, [0, 2]] -= pad_x
-    coords[:, [1, 3]] -= pad_y
-    coords[:, 0:4] /= gain
-    coords[:4] = np.clip(coords[:4], a_min=0, a_max=coords[:4].max())
-    return coords
-
-
-def clip_box(xxyy, img0_shape):
-    xxyy[:, [0, 2]] = np.clip(xxyy[:, [0, 2]], a_min=0, a_max=img0_shape[1])
-    xxyy[:, [1, 3]] = np.clip(xxyy[:, [1, 3]], a_min=0, a_max=img0_shape[0])
-    return xxyy
